requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
lxml==4.9.3
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
urllib3==2.0.6